lang: Build support for dual-language hex files
Detect the printer type by checking the current variant type.

On printers with no xflash (MK2*), generate one hex file for each
additional language file by patching the built-in secondary language
table during the build process

+++ lang/fw-clean.sh
@@ -1,6 +1,14 @@
-#!/bin/sh
+#!/bin/bash
 set -e
+
+# Config
+if [ -z "$CONFIG_OK" ]; then source config.sh; fi
+if [ -z "$CONFIG_OK" -o "$CONFIG_OK" -eq 0 ]; then echo "$(tput setaf 1)Config NG!$(tput sgr0)" >&2; exit 1; fi
 
 # Clean the temporary directory
 TMPDIR=$(dirname "$0")/tmp
 rm -rf "$TMPDIR"
+
+# Remove internationalized firmware files
+rm -f "${OUTHEX_P}${OUTHEX_S}"
+rm -f "${OUTHEX_P}-en_"*"${OUTHEX_S}"
